+++ floodmodeller_api/__init__.py
@@ -6,5 +6,4 @@ from .logs import LF1, LF2
 from .util import read_file
 from .version import __version__
 from .xml2d import XML2D
-from .zzn import ZZN
-from .zzx import ZZX
+from .zz import ZZN, ZZX
